Refresh upon more events
I really want focus change to be one of these events but sadly iridium
changing the focus triggers that event and the state takes around 0.2
seconds to actually be set. This means iridium thinks the focus was
immediately changed back to the original window by the user and so it
tries to resets it back again.

Not sure how to reconcile the separate states. Maybe focus change
shouldn't trigger a refresh of setting the focus.